requirements.txt
beautifulsoup4==4.6.0
bleach==1.5.0
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.3
click==6.7
DateTime==4.2
dill==0.2.7.1
Django==2.0.1
ez-setup==0.9
Flask==0.12.2
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gcloud==0.17.0
geopy==1.11.0
google-api-core==0.1.4
google-auth==1.3.0
google-cloud==0.32.0
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.0
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.0
google-cloud-core==0.28.0
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.0
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.0
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.0
google-cloud-runtimeconfig==0.28.0
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.0
google-cloud-videointelligence==1.0.0
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.8.4
gunicorn==19.7.1
h5py==2.7.1
html5lib==0.9999999
httplib2==0.10.3
idna==2.6
image==1.5.17
itsdangerous==0.24
Jinja2==2.9.6
jws==0.1.3
Keras==2.1.3
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.0
oauth2client==3.0.0
Pillow==5.0.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.5.1
psutil==5.4.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycryptodome==3.4.3
pyreadline==2.1
Pyrebase==3.0.27
python-firebase==1.2
python-jwt==2.0.1
python-resize-image==1.1.11
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-toolbelt==0.7.0
rsa==3.4.2
scipy==1.0.0
six==1.11.0
tensorflow-tensorboard==0.4.0
tf==1.0.0
urllib3==1.22
webcolors==1.7
Werkzeug==0.14.1
zope.interface==4.4.3